+++ tests/AnomalyDetectors/IsolationForestTest.php
@@ -30,7 +30,7 @@ class IsolationForestTest extends TestCase
             1 => new Circle(0., 0., 8., 0.1),
         ], [0.9, 0.1]);
 
-        $this->estimator = new IsolationForest(350, 0.1, 0.3);
+        $this->estimator = new IsolationForest(300, 0.1, 0.4);
     }
 
     public function test_build_detector()